# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/show-allwishlist": {
            "get": {
                "tags": [
                    "Wish list"
                ],
                "summary": "Get all wish lists",
                "operationId": "1273346dc4bbf94b514e68be407e64e8",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error"
                    },
                    "404": {
                        "description": "Wish list not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-wish-list/{id}": {
            "delete": {
                "tags": [
                    "Wish list"
                ],
                "summary": "Delete a wish list by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "e0fdcd1399df7a92f85aaa9e675e1e82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the wish list to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "wish list deleted successfully"
                    },
                    "404": {
                        "description": "wish list not found"
                    },
                    "500": {
                        "description": "Failed to delete wish list"
                    },
                    "400": {
                        "description": "Invalid wish list ID"
                    }
                }
            }
        },
        "/api/user-send-contact": {
            "post": {
                "tags": [
                    "Contacts"
                ],
                "description": "Send an email",
                "operationId": "ac2e6760e97c5b09633a302716b60f00",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Sender's name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Sender's email address",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "john@example.com"
                        }
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "Email subject",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Regarding your inquiry"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Email message content",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Dear Support Team, ..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email sent successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to send email"
                    }
                }
            }
        },
        "/api/admin-reply-contact/{id}": {
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Reply to an email by ID",
                "description": "Reply to an email by ID with a message",
                "operationId": "c46b600b92351c588649bb49c8698e83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the contact to reply to",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Email reply details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "description": "Email reply message content",
                                        "type": "string",
                                        "example": "Your message here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email reply sent successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Contact not found"
                    },
                    "405": {
                        "description": "Invalid request method"
                    },
                    "500": {
                        "description": "Failed to send email reply"
                    }
                }
            }
        },
...
        "/api/search-product/{keyword}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Search product by keyword",
                "operationId": "92465db5cfcf0677b6ac7cee7ca013bd",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "description": "keyword of the product to search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}